+++ src/vs/workbench/contrib/processExplorer/browser/media/processExplorer.css
