+++ src/vs/workbench/workbench.web.main.internal.ts
@@ -166,6 +166,9 @@ import './contrib/splash/browser/splash.contribution.js';
 // Remote Start Entry for the Web
 import './contrib/remote/browser/remoteStartEntry.contribution.js';
 
+// Process Explorer
+import './contrib/processExplorer/browser/processExplorer.web.contribution.js';
+
 //#endregion
